feat: implement patch name and author editing in BoardTopBar and integrate DeckPop component
